// swiyu-ecosystem-portal-ui/src/app/api/generated/model/api-error.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {ApiErrorCode} from './api-error-code';

/**
 * Api error
 */
export interface ApiError {
  errorCode?: ApiErrorCode;
  message?: string;
  traceId?: string;
  additionalDetails?: Array<string>;
}
export namespace ApiError {}