// swiyu-ecosystem-portal-ui/src/app/api/generated/model/api-error-code.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const ApiErrorCode = {
  DataInvalid: 'DATA_INVALID',
  DataInvalidVirusDetected: 'DATA_INVALID_VIRUS_DETECTED',
  ResourceNotFound: 'RESOURCE_NOT_FOUND',
  ResourceForbidden: 'RESOURCE_FORBIDDEN',
  BusinessDataIntegrityViolation: 'BUSINESS_DATA_INTEGRITY_VIOLATION',
  ObjectCountLimitReached: 'OBJECT_COUNT_LIMIT_REACHED',
  InvalidPagination: 'INVALID_PAGINATION',
  IdentifierValidationFailed: 'IDENTIFIER_VALIDATION_FAILED',
  MaxSizeExceeded: 'MAX_SIZE_EXCEEDED',
  DocumentNotFound: 'DOCUMENT_NOT_FOUND'
} as const;
export type ApiErrorCode = (typeof ApiErrorCode)[keyof typeof ApiErrorCode];